refactor: reorganize texture matrix modules
